feat(migrations/postgres): Updated typed in columns to be of type integer

+++ crates/bld_core/migrations/postgres/2021-12-13-183554_create_high_avail_storage/up.sql
@@ -1,7 +1,7 @@
 -- Your SQL goes here
 create table ha_state_machine (
-  id bigint generated always as identity primary key not null,
-  last_applied_log bigint not null,
+  id integer generated always as identity primary key not null,
+  last_applied_log integer not null,
   date_created text default current_timestamp not null,
   date_updated text default current_timestamp not null
 );
@@ -18,9 +18,9 @@ after update on ha_state_machine
 for each row execute procedure ha_state_machine_after_update();
 
 create table ha_client_serial_responses (
-  id bigint primary key not null,
-  state_machine_id bigint not null,
-  serial bigint not null,
+  id integer primary key not null,
+  state_machine_id integer not null,
+  serial integer not null,
   response text,
   date_created text default current_timestamp not null,
   date_updated text default current_timestamp not null,
@@ -39,8 +39,8 @@ after update on ha_client_serial_responses
 for each row execute procedure ha_client_serial_responses_after_update();
 
 create table ha_client_status (
-  id bigint primary key not null,
-  state_machine_id bigint not null,
+  id integer primary key not null,
+  state_machine_id integer not null,
   status text not null,
   date_created text default current_timestamp not null,
   date_updated text default current_timestamp not null,
